Fix NavMap sync of interact objects outside of nav areas
The NavMap distinguishes whether an interact object is outside any
nav area or in a nav area but blocked by a NegZone or NegCircle.

g3dit did not make this distinction, and instead handled interact
object outside any nav area, as if they were not registered in the
NavMap. That resulted in duplicate entries for interact objects.