Improve connection management of Gothic 3 remote control
Depending on the ephemeral port range of the operating system it can
happen that the remote port we try to connect to is assigned as the
ephemeral local port. That results in us connecting to ourselves, and
then requests we send are immediately delivered to us as responses.

Also get rid of the unsent message expire mechanism that was the reason
for the custom built JeroMQ. Instead we just track whether the socket is
connected or not, and only try to send in the former case.